requirements.txt
livekit-agents>=1.0,<2.0
livekit-plugins-silero>=1.0,<2.0
faster-whisper>=1.1.0
python-dotenv>=1.0.0
numpy<2.0
groq>=0.4.0
rank-bm25>=0.2.2
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.15.0
